First version with support for visualizing steps at design time.
This version should work if the steps are defined with inline XAML, e.g.
<wiz:Wizard>
  <wiz:WizardStep>
    <!-- XAML definition of step 1 -->
  </wiz:WizardStep>
  <wiz:WizardStep>
    <!-- XAML definition of step 2 -->
  </wiz:WizardStep>
</wiz:Wizard>

It does not work if the steps are defined in separate UserControls.

The most central change is that Wizard.xaml includes the content of the step using
    <ContentPresenter Grid.Row="0" ContentSource="SelectedItem" />
instead of
    <ContentControl Grid.Row="0"
                    Content="{Binding CurrentStep.Content, RelativeSource={RelativeSource TemplatedParent}}" />

This is the most straight forward way to use the Selector control
(System.Windows.Controls.Primitives.Selector)

Unfortunately, this breaks the previous implementation of the "summary items", which are the dots at the bottom of the wizard, where each dot represents a step in the wizard.

Previously, the summary items where implemented as an ItemsControl containing the WizardSteps. The ControlTemplate of the WizardStep was used to display the "summary item".
This is incompatible with the use of ContentPresenter described above, because it would include the same controls at two positions in the WPF Visual Tree.

To get around this problem, I have added an additional Control object, named WizardStepSummary, and moved most of the functionality of the WizardStep control into the WizardStepSummary control.
There is a one-to-one relationship between the WizardStep and the WizardStepSummary.

In the OnLoaded method in the wizard control, the control builds a list of the WizardStepSummary controls. The ItemsControl for the summary items now binds to the collection of WizardStepSummary controls, instead of the using the Items collection.

The control template for WizardStepSummary is equivalent to the control template for the WizardStep.
I have replaced the control template for the WizardStep with a trivial template contining only
<ContentPresenter Content="{TemplateBinding Content}"/>

There are still two problems:
1. As mentioned above, the design view only works with inline XAML
2. The custom style in RectangleWizardStepStyle.xaml in the WizardDemo does not work. Currently I have no idea why not.