Add support for a Help button, which would usually be positioned on the left.
Of course, a Help button could be defined in the containing window, but I want
the look-and-feel to be exactly like the navigation buttons.
I also want it to be positioned in the same row, with in the area of the wizard.
(I'm not much of a fan of defining a button in the container, but placing it
on top of another control.)

I have tried to implement the Help button in an identical manner to the existing
buttons. In particular this means that the style properties are defined on the
WizardStep control.

In detail:
- Add new properties to the Wizard
  - ShowHelpButton
  - HelpButtonHorizontalAlignment
  - HelpButtonStyle
  - HelpCommand

- Add the full set of button properties to WizardStep for the Help button,
  identical to the Back, Skip and Forward buttons (23 properties).

- Define a HelpButtonStyle in Button.xaml, similar to the other buttons.

- Remove the Width properties from default icon definitions in Button.xaml.
  It is sufficient to specify only the height. For the question mark icon,
  specifying both Width and Height distorts the aspect ratio of the path.

- Add the Help button to the WizardTemplate in Wizard.xaml