[1] Replace all of the Dependency Properties which are still defined on WizardStep with Attached Properties defined on Wizard. These are the properties Visibility, Tooltip, Title and IsEnabled for the four buttons (Help, Back, Skip, Forward).
[2]
Define additional properties (not dependency properties) for binding, without the word "Button",
e.g. BackVisibility instead of BackButtonVisibility, with a fallback logic.
These properties use the property attached to the current step, if it is defined,
or the property attached to the wizard, if it is defined,
or a hard coded default.

Call RaisePropertyChanged on all of these properties when the CurrentStep changes, in Wizard.OnSelectionChanged().

[3]
Remove the property ShowHelpButton. The attached property HelpButtonVisibility should be used instead.