Handle changes to attached properties on steps
There are twenty  properties (Visibility, IsEnalbed, Tooltip and Title on five buttons) which can be specified as attached properties on a step or on the wizard. Attach to change notifications on the attached properties on the current step, and detach we navigate away from the step.

I expect there is a much better way of handling this.